+++ test/components/alg-paper-toggle-button/components.js
@@ -1,0 +1,8 @@
+import '../demo-pages.js';
+import './style.js';
+
+import { AlgPaperToggleButton } from '../../../src/alg-components/paper/alg-paper-toggle-button.js';
+
+document.addEventListener('DOMContentLoaded', () => {
+  document.querySelector('body').style.visibility = '';
+});